Add GPS info component to overview page, closes #96
A new component, 'GpsInfo', has been added to the 'OverviewRoot' vue file. This change enriches the overview page by providing users with GPS-related information such as last recorded position and its accuracy, with an option to update it. Relevant updates have also been made in the 'travelLog' store to accommodate this new component.